move utility functions to a separate file
